separate tutorial for batch
