+++ tutorial/CMakeLists.txt
@@ -128,4 +128,5 @@ if(LIBIGL_TUTORIALS_CHAPTER9)
     igl_add_tutorial(906_TrimWithSolid igl::glfw igl_copyleft::cgal)
     igl_add_tutorial(907_DynamicAABB igl::glfw)
     igl_add_tutorial(908_IntersectionBlockingDecimation igl::glfw igl::predicates)
+    igl_add_tutorial(909_BatchMarchingCubes igl::glfw)
 endif()
